Added the 'stdlib' folder, moved 'View' out of the 'Constant' folder
