+++ language-plutus-core/language-plutus-core.cabal
@@ -33,6 +33,14 @@ library
         Language.PlutusCore.Constant
         Language.PlutusCore.CkMachine
         Language.PlutusCore.TestSupport
+        Language.PlutusCore.View
+        Language.PlutusCore.StdLib.Data.Bool
+        Language.PlutusCore.StdLib.Data.ChurchNat
+        Language.PlutusCore.StdLib.Data.Function
+        Language.PlutusCore.StdLib.Data.List
+        Language.PlutusCore.StdLib.Data.Nat
+        Language.PlutusCore.StdLib.Data.Unit
+        Language.PlutusCore.StdLib.Type
         PlutusPrelude
     build-tools: alex -any, happy >=1.17.1
     hs-source-dirs: src prelude recursion
@@ -42,9 +50,7 @@ library
         Language.PlutusCore.Lexer
         Language.PlutusCore.Lexer.Type
         Language.PlutusCore.Parser
-        Language.PlutusCore.Constant.Prelude
         Language.PlutusCore.Constant.Make
-        Language.PlutusCore.Constant.View
         Language.PlutusCore.Constant.Typed
         Language.PlutusCore.Constant.Apply
         Language.PlutusCore.Renamer
@@ -91,7 +97,6 @@ test-suite language-plutus-core-test
         Evaluation.Constant.Apply
         Evaluation.Constant.Success
         Evaluation.Constant.SuccessFailure
-        Evaluation.Terms
         Evaluation.TypeCheck
         Generators
     default-language: Haskell2010
