+++ language-plutus-core/src/Language/PlutusCore/Constant.hs
@@ -2,8 +2,6 @@ module Language.PlutusCore.Constant
     ( module Export
     ) where
 
-import           Language.PlutusCore.Constant.Prelude as Export
 import           Language.PlutusCore.Constant.Make    as Export
-import           Language.PlutusCore.Constant.View    as Export
 import           Language.PlutusCore.Constant.Typed   as Export
 import           Language.PlutusCore.Constant.Apply   as Export
