+++ language-plutus-core/src/Language/PlutusCore/Constant/Typed.hs
@@ -47,7 +47,6 @@ module Language.PlutusCore.Constant.Typed
     , typedTxHash
     ) where
 
-import           Language.PlutusCore.Constant.Prelude
 import           Language.PlutusCore.Lexer.Type       (BuiltinName (..), TypeBuiltin (..), prettyBytes)
 import           Language.PlutusCore.Name
 import           Language.PlutusCore.Type
