+++ language-plutus-core/src/Language/PlutusCore/View.hs
@@ -1,4 +1,4 @@
-module Language.PlutusCore.Constant.View
+module Language.PlutusCore.View
     ( IterApp(..)
     , TermIterApp
     , PrimIterApp
@@ -9,7 +9,6 @@ module Language.PlutusCore.Constant.View
     , termAsPrimIterApp
     ) where
 
-import           Language.PlutusCore.Constant.Prelude
 import           Language.PlutusCore.Lexer.Type       (BuiltinName (..))
 import           Language.PlutusCore.Type
 import           PlutusPrelude
